+++ app/CHANGELOG.md
@@ -1,3 +1,8 @@
+## What's Changed in app@0.1.1
+* ci: switch to using Xcode 16 on CI by @fortmarek in [#6761](https://github.com/tuist/tuist/pull/6761)
+
+**Full Changelog**: https://github.com/tuist/tuist/compare/app@0.1.0...app@0.1.1
+
 ## What's Changed in app@0.1.0
 * feat: add distribution for the Tuist App by @fortmarek in [#6618](https://github.com/tuist/tuist/pull/6618)